+++ src/Builders/BrandBuilder.php
@@ -4,20 +4,14 @@ namespace App\Builders;
 
 use BBC\ProgrammesPagesService\Domain\Entity\Brand;
 
-class BrandBuilder extends AbstractProgrammeContainerBuilder implements BuilderInterface
+class BrandBuilder extends AbstractProgrammeContainerBuilder
 {
-    protected function __construct()
-    {
-        parent::__construct();
-    }
-
-    /**
-     * Create a builder with default state.
-     * This default state reduce the amount of steps necessary when trying to get a final entity in an specified state
-     */
     public static function default()
     {
-        return new self();
+        $self = new self();
+        $self->withTitle('This is a brand');
+
+        return $self;
     }
 
     /**
